+++ src/codeql.test.ts
@@ -134,7 +134,9 @@ test("downloads and caches explicitly requested bundles that aren't in the toolc
       t.assert(toolcache.find("CodeQL", `0.0.0-${version}`));
       t.is(result.toolsVersion, `0.0.0-${version}`);
       t.is(result.toolsSource, ToolsSource.Download);
-      t.assert(Number.isInteger(result.toolsDownloadDurationMs));
+      t.assert(
+        Number.isInteger(result.toolsDownloadStatusReport?.downloadDurationMs),
+      );
     }
 
     t.is(toolcache.findAllVersions("CodeQL").length, 2);
@@ -162,7 +164,9 @@ test("caches semantically versioned bundles using their semantic version number"
     t.assert(toolcache.find("CodeQL", `2.14.0`));
     t.is(result.toolsVersion, `2.14.0`);
     t.is(result.toolsSource, ToolsSource.Download);
-    t.assert(Number.isInteger(result.toolsDownloadDurationMs));
+    t.assert(
+      Number.isInteger(result.toolsDownloadStatusReport?.downloadDurationMs),
+    );
   });
 });
 
@@ -191,42 +195,36 @@ test("downloads an explicitly requested bundle even if a different version is ca
     t.assert(toolcache.find("CodeQL", "0.0.0-20200610"));
     t.deepEqual(result.toolsVersion, "0.0.0-20200610");
     t.is(result.toolsSource, ToolsSource.Download);
-    t.assert(Number.isInteger(result.toolsDownloadDurationMs));
+    t.assert(
+      Number.isInteger(result.toolsDownloadStatusReport?.downloadDurationMs),
+    );
   });
 });
 
 const EXPLICITLY_REQUESTED_BUNDLE_TEST_CASES = [
   {
-    cliVersion: "2.10.0",
-    expectedToolcacheVersion: "2.10.0-20200610",
+    tagName: "codeql-bundle-2.17.6",
+    expectedToolcacheVersion: "2.17.6",
   },
   {
-    cliVersion: "2.10.0-pre",
-    expectedToolcacheVersion: "0.0.0-20200610",
-  },
-  {
-    cliVersion: "2.10.0+202006100101",
-    expectedToolcacheVersion: "0.0.0-20200610",
+    tagName: "codeql-bundle-20240805",
+    expectedToolcacheVersion: "0.0.0-20240805",
   },
 ];
 
 for (const {
-  cliVersion,
+  tagName,
   expectedToolcacheVersion,
 } of EXPLICITLY_REQUESTED_BUNDLE_TEST_CASES) {
-  test(`caches an explicitly requested bundle containing CLI ${cliVersion} as ${expectedToolcacheVersion}`, async (t) => {
+  test(`caches explicitly requested bundle ${tagName} as ${expectedToolcacheVersion}`, async (t) => {
     await util.withTmpDir(async (tmpDir) => {
       setupActionsVars(tmpDir, tmpDir);
 
       mockApiDetails(SAMPLE_DOTCOM_API_DETAILS);
       sinon.stub(actionsUtil, "isRunningLocalAction").returns(true);
 
-      const releaseApiMock = mockReleaseApi({
-        assetNames: [`cli-version-${cliVersion}.txt`],
-        tagName: "codeql-bundle-20200610",
-      });
       const url = mockBundleDownloadApi({
-        tagName: "codeql-bundle-20200610",
+        tagName,
       });
 
       const result = await codeql.setupCodeQL(
@@ -238,11 +236,12 @@ for (const {
         getRunnerLogger(true),
         false,
       );
-      t.assert(releaseApiMock.isDone(), "Releases API should have been called");
       t.assert(toolcache.find("CodeQL", expectedToolcacheVersion));
-      t.deepEqual(result.toolsVersion, cliVersion);
+      t.deepEqual(result.toolsVersion, expectedToolcacheVersion);
       t.is(result.toolsSource, ToolsSource.Download);
-      t.assert(Number.isInteger(result.toolsDownloadDurationMs));
+      t.assert(
+        Number.isInteger(result.toolsDownloadStatusReport?.downloadDurationMs),
+      );
     });
   });
 }
@@ -277,7 +276,7 @@ for (const toolcacheVersion of [
         );
         t.is(result.toolsVersion, SAMPLE_DEFAULT_CLI_VERSION.cliVersion);
         t.is(result.toolsSource, ToolsSource.Toolcache);
-        t.is(result.toolsDownloadDurationMs, undefined);
+        t.is(result.toolsDownloadStatusReport?.downloadDurationMs, undefined);
       });
     },
   );
@@ -307,7 +306,7 @@ test(`uses a cached bundle when no tools input is given on GHES`, async (t) => {
     );
     t.deepEqual(result.toolsVersion, "0.0.0-20200601");
     t.is(result.toolsSource, ToolsSource.Toolcache);
-    t.is(result.toolsDownloadDurationMs, undefined);
+    t.is(result.toolsDownloadStatusReport?.downloadDurationMs, undefined);
 
     const cachedVersions = toolcache.findAllVersions("CodeQL");
     t.is(cachedVersions.length, 1);
@@ -341,7 +340,9 @@ test(`downloads bundle if only an unpinned version is cached on GHES`, async (t)
     );
     t.deepEqual(result.toolsVersion, defaults.cliVersion);
     t.is(result.toolsSource, ToolsSource.Download);
-    t.assert(Number.isInteger(result.toolsDownloadDurationMs));
+    t.assert(
+      Number.isInteger(result.toolsDownloadStatusReport?.downloadDurationMs),
+    );
 
     const cachedVersions = toolcache.findAllVersions("CodeQL");
     t.is(cachedVersions.length, 2);
@@ -372,7 +373,9 @@ test('downloads bundle if "latest" tools specified but not cached', async (t) =>
     );
     t.deepEqual(result.toolsVersion, defaults.cliVersion);
     t.is(result.toolsSource, ToolsSource.Download);
-    t.assert(Number.isInteger(result.toolsDownloadDurationMs));
+    t.assert(
+      Number.isInteger(result.toolsDownloadStatusReport?.downloadDurationMs),
+    );
 
     const cachedVersions = toolcache.findAllVersions("CodeQL");
     t.is(cachedVersions.length, 2);
@@ -386,7 +389,7 @@ test("bundle URL from another repo is cached as 0.0.0-bundleVersion", async (t) 
     mockApiDetails(SAMPLE_DOTCOM_API_DETAILS);
     sinon.stub(actionsUtil, "isRunningLocalAction").returns(true);
     const releasesApiMock = mockReleaseApi({
-      assetNames: ["cli-version-2.12.6.txt"],
+      assetNames: ["cli-version-2.13.5.txt"],
       tagName: "codeql-bundle-20230203",
     });
     mockBundleDownloadApi({
@@ -407,7 +410,9 @@ test("bundle URL from another repo is cached as 0.0.0-bundleVersion", async (t) 
 
     t.is(result.toolsVersion, "0.0.0-20230203");
     t.is(result.toolsSource, ToolsSource.Download);
-    t.true(Number.isInteger(result.toolsDownloadDurationMs));
+    t.true(
+      Number.isInteger(result.toolsDownloadStatusReport?.downloadDurationMs),
+    );
 
     const cachedVersions = toolcache.findAllVersions("CodeQL");
     t.is(cachedVersions.length, 1);
@@ -689,7 +694,7 @@ test("passes a code scanning config AND qlconfig to the CLI", async (t: Executio
   await util.withTmpDir(async (tempDir) => {
     const runnerConstructorStub = stubToolRunnerConstructor();
     const codeqlObject = await codeql.getCodeQLForTesting();
-    sinon.stub(codeqlObject, "getVersion").resolves(makeVersionInfo("2.12.6"));
+    sinon.stub(codeqlObject, "getVersion").resolves(makeVersionInfo("2.17.6"));
 
     await codeqlObject.databaseInitCluster(
       { ...stubConfig, tempDir },
@@ -718,7 +723,7 @@ test("does not pass a qlconfig to the CLI when it is undefined", async (t: Execu
   await util.withTmpDir(async (tempDir) => {
     const runnerConstructorStub = stubToolRunnerConstructor();
     const codeqlObject = await codeql.getCodeQLForTesting();
-    sinon.stub(codeqlObject, "getVersion").resolves(makeVersionInfo("2.12.6"));
+    sinon.stub(codeqlObject, "getVersion").resolves(makeVersionInfo("2.17.6"));
 
     await codeqlObject.databaseInitCluster(
       { ...stubConfig, tempDir },
@@ -759,7 +764,7 @@ const NEW_ANALYSIS_SUMMARY_TEST_CASES = [
     codeqlVersion: makeVersionInfo("2.15.0"),
     githubVersion: {
       type: util.GitHubVariant.GHES,
-      version: "3.9.0",
+      version: "3.10.0",
     },
     flagPassed: true,
     negativeFlagPassed: false,
@@ -804,7 +809,6 @@ for (const {
       "",
       Object.assign({}, stubConfig, { gitHubVersion: githubVersion }),
       createFeatures([]),
-      getRunnerLogger(true),
     );
     const actualArgs = runnerConstructorStub.firstCall.args[1] as string[];
     t.is(
@@ -822,32 +826,6 @@ for (const {
   });
 }
 
-test("database finalize does not override no code found error on CodeQL 2.12.6", async (t) => {
-  const cliMessage =
-    "CodeQL did not detect any code written in languages supported by CodeQL. Review our troubleshooting guide at " +
-    "https://gh.io/troubleshooting-code-scanning/no-source-code-seen-during-build.";
-  stubToolRunnerConstructor(32, cliMessage);
-  const codeqlObject = await codeql.getCodeQLForTesting();
-  sinon.stub(codeqlObject, "getVersion").resolves(makeVersionInfo("2.12.6"));
-  // safeWhich throws because of the test CodeQL object.
-  sinon.stub(safeWhich, "safeWhich").resolves("");
-
-  await t.throwsAsync(
-    async () =>
-      await codeqlObject.finalizeDatabase(
-        "db",
-        "--threads=2",
-        "--ram=2048",
-        false,
-      ),
-    {
-      message:
-        'Encountered a fatal error while running "codeql-for-testing database finalize --finalize-dataset --threads=2 --ram=2048 db". ' +
-        `Exit code was 32 and last log line was: ${cliMessage} See the logs for more details.`,
-    },
-  );
-});
-
 test("runTool summarizes several fatal errors", async (t) => {
   const heapError =
     "A fatal error occurred: Evaluator heap must be at least 384.00 MiB";
@@ -858,7 +836,7 @@ test("runTool summarizes several fatal errors", async (t) => {
     `${heapError}\n${datasetImportError}.`;
   stubToolRunnerConstructor(32, cliStderr);
   const codeqlObject = await codeql.getCodeQLForTesting();
-  sinon.stub(codeqlObject, "getVersion").resolves(makeVersionInfo("2.12.6"));
+  sinon.stub(codeqlObject, "getVersion").resolves(makeVersionInfo("2.17.6"));
   // safeWhich throws because of the test CodeQL object.
   sinon.stub(safeWhich, "safeWhich").resolves("");
 
@@ -899,7 +877,7 @@ test("runTool summarizes autobuilder errors", async (t) => {
   `;
   stubToolRunnerConstructor(1, stderr);
   const codeqlObject = await codeql.getCodeQLForTesting();
-  sinon.stub(codeqlObject, "getVersion").resolves(makeVersionInfo("2.12.6"));
+  sinon.stub(codeqlObject, "getVersion").resolves(makeVersionInfo("2.17.6"));
   sinon.stub(codeqlObject, "resolveExtractor").resolves("/path/to/extractor");
   // safeWhich throws because of the test CodeQL object.
   sinon.stub(safeWhich, "safeWhich").resolves("");
@@ -926,7 +904,7 @@ test("runTool truncates long autobuilder errors", async (t) => {
   ).join("\n");
   stubToolRunnerConstructor(1, stderr);
   const codeqlObject = await codeql.getCodeQLForTesting();
-  sinon.stub(codeqlObject, "getVersion").resolves(makeVersionInfo("2.12.6"));
+  sinon.stub(codeqlObject, "getVersion").resolves(makeVersionInfo("2.17.6"));
   sinon.stub(codeqlObject, "resolveExtractor").resolves("/path/to/extractor");
   // safeWhich throws because of the test CodeQL object.
   sinon.stub(safeWhich, "safeWhich").resolves("");
@@ -954,7 +932,7 @@ test("runTool recognizes fatal internal errors", async (t) => {
     Severe disk cache trouble (corruption or out of space) at /home/runner/work/_temp/codeql_databases/go/db-go/default/cache/pages/28/33.pack: Failed to write item to disk`;
   stubToolRunnerConstructor(1, stderr);
   const codeqlObject = await codeql.getCodeQLForTesting();
-  sinon.stub(codeqlObject, "getVersion").resolves(makeVersionInfo("2.12.6"));
+  sinon.stub(codeqlObject, "getVersion").resolves(makeVersionInfo("2.17.6"));
   sinon.stub(codeqlObject, "resolveExtractor").resolves("/path/to/extractor");
   // safeWhich throws because of the test CodeQL object.
   sinon.stub(safeWhich, "safeWhich").resolves("");
@@ -964,7 +942,7 @@ test("runTool recognizes fatal internal errors", async (t) => {
       await codeqlObject.databaseRunQueries(stubConfig.dbLocation, []),
     {
       instanceOf: CommandInvocationError,
-      message: `Encountered a fatal error while running "codeql-for-testing database run-queries  --expect-discarded-cache --min-disk-free=1024 -v". Exit code was 1 and error was: Oops! A fatal internal error occurred. Details:
+      message: `Encountered a fatal error while running "codeql-for-testing database run-queries  --expect-discarded-cache --min-disk-free=1024 -v --intra-layer-parallelism". Exit code was 1 and error was: Oops! A fatal internal error occurred. Details:
     com.semmle.util.exception.CatastrophicError: An error occurred while evaluating ControlFlowGraph::ControlFlow::Root.isRootOf/1#dispred#f610e6ed/2@86282cc8
     Severe disk cache trouble (corruption or out of space) at /home/runner/work/_temp/codeql_databases/go/db-go/default/cache/pages/28/33.pack: Failed to write item to disk. See the logs for more details.`,
     },
@@ -975,7 +953,7 @@ test("runTool outputs last line of stderr if fatal error could not be found", as
   const cliStderr = "line1\nline2\nline3\nline4\nline5";
   stubToolRunnerConstructor(32, cliStderr);
   const codeqlObject = await codeql.getCodeQLForTesting();
-  sinon.stub(codeqlObject, "getVersion").resolves(makeVersionInfo("2.12.6"));
+  sinon.stub(codeqlObject, "getVersion").resolves(makeVersionInfo("2.17.6"));
   // safeWhich throws because of the test CodeQL object.
   sinon.stub(safeWhich, "safeWhich").resolves("");
 
@@ -1000,7 +978,7 @@ test("runTool outputs last line of stderr if fatal error could not be found", as
 test("Avoids duplicating --overwrite flag if specified in CODEQL_ACTION_EXTRA_OPTIONS", async (t) => {
   const runnerConstructorStub = stubToolRunnerConstructor();
   const codeqlObject = await codeql.getCodeQLForTesting();
-  sinon.stub(codeqlObject, "getVersion").resolves(makeVersionInfo("2.12.6"));
+  sinon.stub(codeqlObject, "getVersion").resolves(makeVersionInfo("2.17.6"));
   // safeWhich throws because of the test CodeQL object.
   sinon.stub(safeWhich, "safeWhich").resolves("");