+++ src/start-proxy-action.ts
@@ -156,6 +156,7 @@ async function startProxy(binPath: string, config: ProxyConfig, logFilePath: str
       await util.delay(1000);
     }
     if (subprocessError) {
+      // eslint-disable-next-line @typescript-eslint/only-throw-error
       throw subprocessError;
     }
     logger.info(`Proxy started on ${host}:${port}`);